Slowly getting there with WebSockets.
Data is being sent from the client to the server and the server is
generating zombies and sending that data to the client, but the
bearing() function isn't reading the latitude or longitude properties
of the zombie objects which means we can't draw the zombies properly on
the screen.  There's some error happening when switching the generated
data to JSON or back.